CX-114: Receive values from BK calls B_GEOMETRY_PIXEL and B_GEOMETRY_TISSUE

+++ source/plugins/org.custusx.core.openigtlink3/testing/cxtestOpenIGTLinkIO.cpp
@@ -75,7 +75,7 @@ void checkIfConnected(igtlio::LogicPointer logic)
 		}
 	}
 
-	REQUIRE(connector->GetState() == igtlio::Connector::STATE_CONNECTED);
+	REQUIRE(connector->GetState() == static_cast<int>(igtlio::Connector::STATE_CONNECTED));
 }
 
 void tryToReceiveEvents(igtlio::LogicPointer logic, Receiver &receiver)
